@@ -1,6 +1,7 @@
 package com.example.ramirez;
 
 import android.content.Intent;
+import android.os.Build;
 import android.os.Bundle;
 import android.os.StrictMode;
 import android.view.View;
@@ -9,41 +10,68 @@ import android.widget.ImageView;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import androidx.annotation.RequiresApi;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.example.ramirez.adapter.PhotoRecyclerViewAdapter;
-import com.example.ramirez.dao.PostDAO;
 import com.example.ramirez.helpers.FirebaseHelper;
 import com.example.ramirez.helpers.RecyclerItemClickListener;
 import com.example.ramirez.helpers.SessionManager;
-import com.example.ramirez.helpers.UsersService;
+import com.example.ramirez.services.PostService;
+import com.example.ramirez.services.UsersService;
 import com.example.ramirez.model.Photographer;
 import com.example.ramirez.model.Post;
+import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.auth.FirebaseUser;
 import com.google.firebase.storage.StorageReference;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 
 public class ProfileUserActivity extends AppCompatActivity {
     private RecyclerView postRecyclerView;
-    private List<Post> posts;
+    private List<Post> posts = new ArrayList<>();
     private PhotoRecyclerViewAdapter adapter;
 
+    @RequiresApi(api = Build.VERSION_CODES.N)
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         Objects.requireNonNull(getSupportActionBar()).hide();
         setContentView(R.layout.activity_profile_photo_user);
 
+        SessionManager sessionManager = new SessionManager(this);
+
+        FirebaseAuth mAuth = FirebaseAuth.getInstance();
+        FirebaseUser user = mAuth.getCurrentUser();
+
+        if (user != null) {
+            // do your stuff
+            try {
+                getUserPosts(sessionManager);
+            } catch (InterruptedException e) {
+                e.printStackTrace();
+            }
+        } else {
+            mAuth.signInAnonymously().addOnSuccessListener(this, authResult -> {
+                // do your stuff
+                try {
+                    getUserPosts(sessionManager);
+                } catch (InterruptedException e) {
+                    e.printStackTrace();
+                }
+            })
+            .addOnFailureListener(this, exception -> exception.printStackTrace());
+        }
+
         this.postRecyclerView = findViewById(R.id.listaDePostagem);
-        this.posts = PostDAO.getInstance(getApplicationContext()).getPosts();
 
         StrictMode.ThreadPolicy policy = new StrictMode.ThreadPolicy.Builder().permitAll().build();
         StrictMode.setThreadPolicy(policy);
 
-        SessionManager sessionManager = new SessionManager(this);
         UsersService usersService = UsersService.getInstance(sessionManager);
 
         Photographer currentPhotographer = usersService.getPhotographer(sessionManager.fetchUserId());
@@ -62,8 +90,49 @@ public class ProfileUserActivity extends AppCompatActivity {
         String viewsMessage = currentPhotographer.getViews() + " visualizações";
         userViews.setText(viewsMessage);
 
-        this.adapter = new PhotoRecyclerViewAdapter(this.posts);
+        ImageView editProfileButton = findViewById(R.id.editProfileButton);
+        ImageView publishNewPhoto = findViewById(R.id.publishPostButton);
 
+        editProfileButton.setOnClickListener(v -> {
+            Intent intent = new Intent(getApplicationContext(), EditProfileActivity.class);
+            intent.putExtra("EDIT_PROFILE_ID", sessionManager.fetchUserId());
+            startActivity(intent);
+        });
+
+        publishNewPhoto.setOnClickListener(v -> {
+            Intent intent = new Intent(getApplicationContext(), PostPhotoActivity.class);
+            startActivity(intent);
+        });
+
+    }
+
+
+    @RequiresApi(api = Build.VERSION_CODES.N)
+    public void getUserPosts(SessionManager sessionManager) throws InterruptedException {
+        PostService postService = PostService.getInstance(sessionManager);
+        StorageReference storageRef = FirebaseHelper.getStorageReference();
+        ArrayList<Post> postList =  postService.getPostsOfCurrentUser();
+
+        for (int i = 0; i < postList.size(); i++) {
+            int finalI = i;
+            storageRef.child(postList.get(i).getImage()).getDownloadUrl()
+                .addOnSuccessListener(uri -> {
+                    postList.get(finalI).setImageUri(uri.toString());
+                    this.posts.add(postList.get(finalI));
+                    if (finalI == postList.size() - 1) {
+                        buildAdapter();
+                    }
+                })
+                .addOnFailureListener(e -> e.printStackTrace());
+        }
+    }
+
+    public void buildAdapter() {
+        if (this.posts.isEmpty()) {
+            return;
+        }
+
+        this.adapter = new PhotoRecyclerViewAdapter(this.posts, this);
         RecyclerView.LayoutManager layoutManager = new LinearLayoutManager(getApplicationContext());
         postRecyclerView.setLayoutManager(layoutManager);
         postRecyclerView.hasFixedSize();
@@ -94,19 +163,5 @@ public class ProfileUserActivity extends AppCompatActivity {
                         }
                 )
         );
-
-        ImageView editProfileButton = findViewById(R.id.editProfileButton);
-        ImageView publishNewPhoto = findViewById(R.id.publishPostButton);
-
-        editProfileButton.setOnClickListener(v -> {
-            Intent intent = new Intent(ProfileUserActivity.this, EditProfileActivity.class);
-            intent.putExtra("EDIT_PROFILE_ID", sessionManager.fetchUserId());
-            startActivity(intent);
-        });
-
-        publishNewPhoto.setOnClickListener(v -> {
-            Intent intent = new Intent(ProfileUserActivity.this, PostPhotoActivity.class);
-            startActivity(intent);
-        });
     }
 }
