@@ -1,6 +1,6 @@
 package com.example.ramirez.model;
 
-import android.graphics.Bitmap;
+import android.net.Uri;
 
 import java.util.ArrayList;
 
@@ -8,15 +8,16 @@ public class Post {
     private String id;
     private String name;
     private Float price;
-    private Bitmap image;
+    private String image;
     private ArrayList<Comment> comments;
+    private String imageUri;
 
-    public Post(String id, String name, Float price, Bitmap image, ArrayList<Comment> comments) {
+    public Post(String id, String name, Float price, String image) {
         this.id = id;
         this.name = name;
         this.price = price;
         this.image = image;
-        this.comments = comments;
+        this.imageUri = null;
     }
 
     public String getId() {
@@ -31,8 +32,12 @@ public class Post {
         return this.price;
     }
 
-    public Bitmap getImage() { return this.image; }
+    public String getImage() { return this.image; }
 
     public ArrayList<Comment> getComments() { return this.comments; }
+
+    public String getImageUri() { return this.imageUri; }
+
+    public void setImageUri(String uri) { this.imageUri = uri; }
 
 }
