@@ -1,7 +1,8 @@
-package com.example.ramirez.helpers;
+package com.example.ramirez.services;
 
 import androidx.annotation.NonNull;
 
+import com.example.ramirez.helpers.SessionManager;
 import com.example.ramirez.model.Photographer;
 
 import org.json.JSONArray;
@@ -9,13 +10,11 @@ import org.json.JSONException;
 import org.json.JSONObject;
 
 import java.io.IOException;
-import java.net.URI;
 import java.util.ArrayList;
 import java.util.List;
 
 import okhttp3.Call;
 import okhttp3.Callback;
-import okhttp3.HttpUrl;
 import okhttp3.OkHttpClient;
 import okhttp3.Request;
 import okhttp3.Response;
